example: libssh2_channel_{read,write} fix read/written sizes

+++ example/ssh2_echo.c
@@ -232,11 +232,11 @@ int main(int argc, char *argv[])
     else {
         LIBSSH2_POLLFD *fds = NULL;
         int running = 1;
-        int bufsize = BUFSIZE;
+        ssize_t bufsize = BUFSIZE;
         char buffer[BUFSIZE];
-        int totsize = 1500000;
-        int totwritten = 0;
-        int totread = 0;
+        ssize_t totsize = 1500000;
+        ssize_t totwritten = 0;
+        ssize_t totread = 0;
         int rereads = 0;
         int rewrites = 0;
         int i;
@@ -262,7 +262,8 @@ int main(int argc, char *argv[])
                 continue;
 
             if(fds[0].revents & LIBSSH2_POLLFD_POLLIN) {
-                int n = libssh2_channel_read(channel, buffer, sizeof(buffer));
+                ssize_t n = libssh2_channel_read(channel,
+                                                 buffer, sizeof(buffer));
                 act++;
 
                 if(n == LIBSSH2_ERROR_EAGAIN) {
@@ -275,7 +276,7 @@ int main(int argc, char *argv[])
                 }
                 else {
                     totread += n;
-                    fprintf(stderr, "read %d bytes (%d in total)\n",
+                    fprintf(stderr, "read %zd bytes (%zd in total)\n",
                             n, totread);
                 }
             }
@@ -285,9 +286,10 @@ int main(int argc, char *argv[])
 
                 if(totwritten < totsize) {
                     /* we have not written all data yet */
-                    int left = totsize - totwritten;
-                    int size = (left < bufsize) ? left : bufsize;
-                    int n = libssh2_channel_write_ex(channel, 0, buffer, size);
+                    ssize_t left = totsize - totwritten;
+                    ssize_t size = (left < bufsize) ? left : bufsize;
+                    ssize_t n = libssh2_channel_write_ex(channel, 0,
+                                                         buffer, size);
 
                     if(n == LIBSSH2_ERROR_EAGAIN) {
                         rewrites++;
@@ -299,7 +301,7 @@ int main(int argc, char *argv[])
                     }
                     else {
                         totwritten += n;
-                        fprintf(stderr, "wrote %d bytes (%d in total)",
+                        fprintf(stderr, "wrote %zd bytes (%zd in total)",
                                 n, totwritten);
                         if(left >= bufsize && n != bufsize) {
                             fprintf(stderr, " PARTIAL");
@@ -348,12 +350,12 @@ int main(int argc, char *argv[])
         libssh2_channel_free(channel);
         channel = NULL;
 
-        fprintf(stderr, "\nrereads: %d rewrites: %d totwritten %d\n",
+        fprintf(stderr, "\nrereads: %d rewrites: %d totwritten %zd\n",
                 rereads, rewrites, totwritten);
 
         if(totwritten != totread) {
-            fprintf(stderr, "\n*** FAIL bytes written: %d bytes "
-                    "read: %d ***\n", totwritten, totread);
+            fprintf(stderr, "\n*** FAIL bytes written: %zd bytes "
+                    "read: %zd ***\n", totwritten, totread);
             exit(1);
         }
     }
